jinaai-v3/rag-api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-multipart==0.0.9

qdrant-client==1.11.1

torch==2.4.1
numpy==2.0.2

transformers==4.46.3
sentence-transformers==3.2.0
huggingface_hub==0.26.2
einops==0.8.0

pypdf==4.3.1
python-docx==1.1.2

pytesseract==0.3.10
Pillow==10.4.0
pdf2image==1.17.0